requirements.txt
streamlit
pytesseract
pdf2image
Pillow
selenium
webdriver-manager
openai>=1.0.0